Matplot++ is not needed any more

+++ Lab/PDE/cpp/src/H-1/H-1.cpp
@@ -4,7 +4,7 @@
 #include <iostream>
 #include <functional>
 
-static void biharmonicEvolveExplicitEulerPeriodic(std::function<double(double)> u_initial, double a, double b, unsigned int J, unsigned int T, unsigned int M, double** &u, std::vector<double> &x);
+static void biharmonicEvolveExplicitEulerPeriodic(std::function<double(double)> u_initial, double a, double b, unsigned int J, double T, unsigned int M, double** &u, std::vector<double> &x);
 
 
 int main()
@@ -23,7 +23,7 @@ int main()
     return 0;
 }
 
-static void biharmonicEvolveExplicitEulerPeriodic(std::function<double(double)> u_initial, double a, double b, unsigned int J, unsigned int T, unsigned int M, double** &u, std::vector<double> &x)
+static void biharmonicEvolveExplicitEulerPeriodic(std::function<double(double)> u_initial, double a, double b, unsigned int J, double T, unsigned int M, double** &u, std::vector<double> &x)
 {
     double deltaT = (double) T / M;
     double deltaX = (b - a) / J;
